+++ ext.py
@@ -96,6 +96,7 @@ class Popup(html5.Div):
 		""")
 
 		self.appendChild = self.popupBody.appendChild
+		self.fromHTML = lambda *args, **kwargs: self.popupBody.fromHTML(*args, **kwargs) if kwargs.get("bindTo") else self.popupBody.fromHTML(bindTo=self, *args, **kwargs)
 
 		self["class"] = "popup popup--center is-active"
 		if className:
@@ -159,41 +160,52 @@ class InputDialog(Popup):
 				 	successLbl="OK", abortLbl="Cancel", placeholder="", *args, **kwargs):
 
 		super().__init__(*args, **kwargs)
-		self["class"].append("popup--inputdialog")
+		self.addClass("popup--inputdialog")
+
+		self.sinkEvent("onKeyDown", "onKeyUp")
 
 		self.successHandler = successHandler
 		self.abortHandler = abortHandler
 
-		self.inputGroup = html5.Div()
-		self.inputGroup.addClass("input-group")
-		self.popupBody.appendChild(self.inputGroup)
+		self.fromHTML(
+			"""
+			<div class="input-group">
+				<label class="label">
+					{{text}}
+					<input class="input" [name]="inputElem" value="{{value}}" placeholder="{{placeholder}}" />
+				</label>
+			</div>
+			""",
+			vars={
+				"text": text,
+				"value": value,
+				"placeholder": placeholder
+			}
+		)
 
-		inputLabel = html5.Label()
-		inputLabel.addClass("label")
-		inputLabel.element.innerHTML = text
+		# Cancel
+		self.popupFoot.appendChild(Button(abortLbl, self.onCancel, className="btn--cancel btn--danger"))
 
-		self.inputGroup.appendChild(inputLabel)
-		self.inputElem = html5.Input()
-		self.inputElem.addClass("input")
-		self.inputElem["type"] = "text"
-		self.inputElem["value"] = value
-		self.inputElem["placeholder"] = placeholder
+		# Okay
+		self.okayBtn = Button(successLbl, self.onOkay, className="btn--okay btn--primary")
+		if not value:
+			self.okayBtn.disable()
 
-		self.inputGroup.appendChild(self.inputElem)
-		cancelBtn = Button(abortLbl, self.onCancel, className="btn--cancel btn--danger")
-		self.popupFoot.appendChild(cancelBtn)
+		self.popupFoot.appendChild(self.okayBtn)
 
-		okayBtn = Button(successLbl, self.onOkay)
-		okayBtn["class"].append("btn--okay btn--primary")
-		self.popupFoot.appendChild(okayBtn)
-		self.sinkEvent("onKeyDown")
 		self.inputElem.focus()
 
 	def onKeyDown(self, event):
-		if html5.isReturn(event):
+		if html5.isReturn(event) and self.inputElem["value"]:
 			event.stopPropagation()
 			event.preventDefault()
 			self.onOkay()
+
+	def onKeyUp(self, event):
+		if self.inputElem["value"]:
+			self.okayBtn.enable()
+		else:
+			self.okayBtn.disable()
 
 	def onDocumentKeyDown(self, event):
 		if html5.isEscape(event):